requirements.txt
Flask
python-dotenv
gunicorn
requests
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Login